Move processors to modeling file
